feat(ao3-helpers): add support for keyboard navigation between works in a series
When viewing a work at is part of only one series, you can press the left arrow on the first chapter to navigate to the previous work in the series. Likewise, you can navigate to the next work in the series from the last chapter with the right arrow key.